update proftpd: add new files
git-svn-id: svn://svn.dd-wrt.com/DD-WRT@44069 52c4871e-980c-0410-b1e0-e73912ce01f8

+++ src/router/proftpd/contrib/mod_sftp/blacklist.c
@@ -110,7 +110,7 @@ static int validate_blacklist(int fd, unsigned int *bytes,
 
   *shift = (hdr.shift[0] << 8) + hdr.shift[1];
 
-  expected = sizeof(hdr) + 0x20000 + (*records) * (*bytes);
+  expected = sizeof(hdr) + 0x20000 + ((size_t) (*records) * (*bytes));
   if (st.st_size != (off_t) expected) {
     pr_trace_msg(trace_channel, 4,
       "unexpected SFTPKeyBlacklist '%s' file size: expected %lu, found %lu",
